+++ packages/multitest/tests/proxy-governance-owned.rs
@@ -133,7 +133,7 @@ fn integration_test() {
             &nois_proxy_governance_owned::msg::InstantiateMsg {
                 manager: Some("manager".to_string()),
                 prices: vec![Coin::new(1_000_000, "unoisx")],
-                test_mode: false,
+                test_mode: None,
                 callback_gas_limit: 500_000,
                 mode: nois_proxy_governance_owned::state::OperationalMode::Funded {},
                 allowlist_enabled: None,